backend/requirements.txt
dotenv-python
fastapi[standard]
openai
scikit-learn
numpy
tqdm
uvicorn